requirements.txt
bencode==1.0
mysql-connector-python==2.0.2
python-libtorrent==1.1.1
wxPython==3.0.1.1
wxPython-common==3.0.1.1
